Nikhil_porfoliopersonal-main/project_copy/Studygroup/backend/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Flask-Cors==3.0.10
Flask-SocketIO==5.1.1
python-dotenv==0.19.0
openai==0.27.0
Pillow==10.2.0
pytesseract==0.3.8
python-jose==3.3.0
bcrypt==3.2.0
SQLAlchemy==1.4.23
PyJWT==2.8.0 